# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tobfa's API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/authentication/authenticate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to systems and return the token and inits data.",
                "description": "Do Login",
                "operationId": "23bc458319f733333c2453412dc52c70",
                "parameters": [
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Login Username & Password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "me@tobfa.id"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                },
                                "remember_me": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/authentication/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to systems and return the token and inits data.",
                "description": "Do Logout",
                "operationId": "561148bb8ca22386468e54de95a939b3",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/businesses": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Get business data",
                "description": "Get Business",
                "operationId": "1f79cf73e62dc5a8b542c9e6be704acc",
                "responses": {
                    "200": {
                        "description": "successful get business"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Post business data",
                "description": "Post Business",
                "operationId": "eac1732d2e6210609ae7f4d424dafdd3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Business"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful Post business"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/businesses/{id}": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Find business data",
                "description": "Find Business",
                "operationId": "5e9f4f497b820add110b20cbbdd40a06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful get business"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Put business data",
                "description": "Put Business",
                "operationId": "34dbe9672944022d387c805f7c0fe353",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Business"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful Put business"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Delete business data",
                "description": "Delete Business",
                "operationId": "ba3204b8b7a0eb0ea0ec28eee619b42a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful delete business"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/businessCategories": {
            "get": {
                "tags": [
                    "Business Categories"
                ],
                "summary": "Get business category data",
                "description": "Get Business category",
                "operationId": "692e1f661418b2be03149226e614b670",
                "responses": {
                    "200": {
                        "description": "successful get business category"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/inits": {
            "get": {
                "tags": [
                    "Init"
                ],
                "summary": "Get init data",
                "description": "Get Init",
                "operationId": "4c05e29c3ff31208d40feffed0180e1a",
                "responses": {
                    "200": {
                        "description": "successful get init"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction data",
                "description": "Get transaction",
                "operationId": "9742523dbd7d1b65e62d315300ba7256",
                "responses": {
                    "200": {
                        "description": "successful get transaction item"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Post transaction data",
                "description": "Post Transactions",
                "operationId": "d704a565af6760412bb4a3aefa71af9f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transaction"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful Post transaction"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Put transaction data",
                "description": "Put Transaction",
                "operationId": "9042a6ab13891bd69f6ac5fae7507f52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transaction"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful Put transaction"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete transaction data",
                "description": "Delete Transaction",
                "operationId": "377707e0473c0fd5d0c69b0531de107d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful delete transaction"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/transactionItems": {
            "get": {
                "tags": [
                    "Transaction Items"
                ],
                "summary": "Get transaction item data",
                "description": "Get transaction item",
                "operationId": "fd1268e397714eafdf3975cbf98c5656",
                "responses": {
                    "200": {
                        "description": "successful get transaction item"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/transactionTypes": {
            "get": {
                "tags": [
                    "Transaction Types"
                ],
                "summary": "Get transaction type data",
                "description": "Get transaction type",
                "operationId": "40f9f41ce2a6fddf60cbf0d28885c58e",
                "responses": {
                    "200": {
                        "description": "successful get transaction type"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "description": "Create User",
                "operationId": "181038dd23c1a5f2bbdcff1a0b9cb2f5",
                "parameters": [
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "me"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "me@tobfa.id"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Business": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Usaha 1"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "business_category_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "properties": {
                    "business_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "transaction_type_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "transaction_item_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "total": {
                        "type": "integer",
                        "example": "1000"
                    },
                    "multiplier": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}